[RCP-346] :recycle: Navigation과 홈화면, 마이페이지의 로딩 관계를 개선
