+++ app/src/main/java/com/recipia/aos/ui/components/TopAppBar.kt
@@ -50,69 +50,5 @@ fun TopAppBar(
     bookmarkViewModel: BookMarkViewModel,
 ) {
 
-    val scrollBehavior = TopAppBarDefaults.pinnedScrollBehavior(rememberTopAppBarState())
-    var menuExpanded by remember { mutableStateOf(false) } // 드롭다운 메뉴 상태
 
-    Scaffold(
-        modifier = Modifier
-            .nestedScroll(scrollBehavior.nestedScrollConnection)
-            .background(Color.White),
-        topBar = {
-            CenterAlignedTopAppBar(
-                modifier = Modifier.background(Color.White), // 여기에 배경색을 하얀색으로 설정,
-                colors = TopAppBarDefaults.topAppBarColors(
-                    containerColor = Color.Transparent, // TopAppBar 배경을 투명하게 설정
-                    titleContentColor = MaterialTheme.colorScheme.onPrimaryContainer
-                ),
-                title = {},
-                navigationIcon = {
-                    IconButton(onClick = { navController.navigate("categories") }) {
-                        Icon(
-                            imageVector = Icons.Filled.Menu,
-                            contentDescription = "메뉴"
-                        )
-                    }
-                },
-                actions = {
-                    // 검색 아이콘 추가
-                    IconButton(onClick = { navController.navigate("searchScreen") }) {
-                        Icon(Icons.Filled.Search, contentDescription = "검색")
-                    }
-
-                    IconButton(onClick = { menuExpanded = true }) {
-                        Icon(
-                            imageVector = Icons.Filled.MoreVert,
-                            contentDescription = "더보기"
-                        )
-                    }
-
-                    CustomDropdownMenu(
-                        expanded = menuExpanded,
-                        onDismissRequest = { menuExpanded = false }
-                    ) {
-                        // 드롭다운 메뉴 아이템들
-                        DropdownMenuItem(
-                            text = { Text("신고하기") },
-                            onClick = { /* 수정 처리 */ }
-                        )
-                        DropdownMenuItem(
-                            text = { Text("로그아웃") },
-                            onClick = { /* 설정 처리 */ }
-                        )
-                        DropdownMenuItem(
-                            text = { Text("로그아웃") },
-                            onClick = { /* 설정 처리 */ }
-                        )
-                        DropdownMenuItem(
-                            text = { Text("피드백 보내기") },
-                            onClick = { /* 피드백 처리 */ }
-                        )
-                    }
-                },
-                scrollBehavior = scrollBehavior
-            )
-        }
-    ) { innerPadding ->
-        HomeScreen(navController, viewModel, bookmarkViewModel, innerPadding)
-    }
 }
