+++ app/src/main/java/com/recipia/aos/ui/model/factory/MyViewModelFactory.kt
@@ -1,14 +1,17 @@
 package com.recipia.aos.ui.model.factory
 
-import com.recipia.aos.ui.model.login.LoginViewModel
-import android.content.Context
+import TokenManager
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import com.recipia.aos.ui.model.login.LoginViewModel
 
-class MyViewModelFactory(private val context: Context) : ViewModelProvider.Factory {
+class MyViewModelFactory(
+    private val tokenManager: TokenManager
+) : ViewModelProvider.Factory {
+
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(LoginViewModel::class.java)) {
-            return LoginViewModel(context) as T
+            return LoginViewModel(tokenManager) as T
         }
         throw IllegalArgumentException("Unknown ViewModel class")
     }
